// src/types/movie.ts
// Interface automatically generated by schemas-to-ts

import { Project } from '@/types/project';
import { Project_Plain } from '@/types/project';
import { AdminPanelRelationPropertyModification } from '@/types/AdminPanelRelationPropertyModification';

export interface Movie {
  id: number;
  attributes: {
    createdAt: Date;
    updatedAt: Date;
    publishedAt?: Date;
    Movie: string;
    Description: string;
    ProductionYear: number;
    Links?: any;
    projects: { data: Project[] };
    Slug: string;
  };
}
export interface Movie_Plain {
  id: number;
  createdAt: Date;
  updatedAt: Date;
  publishedAt?: Date;
  Movie: string;
  Description: string;
  ProductionYear: number;
  Links?: any;
  projects: Project_Plain[];
  Slug: string;
}

export interface Movie_NoRelations {
  id: number;
  createdAt: Date;
  updatedAt: Date;
  publishedAt?: Date;
  Movie: string;
  Description: string;
  ProductionYear: number;
  Links?: any;
  projects: number[];
  Slug: string;
}

export interface Movie_AdminPanelLifeCycle {
  id: number;
  createdAt: Date;
  updatedAt: Date;
  publishedAt?: Date;
  Movie: string;
  Description: string;
  ProductionYear: number;
  Links?: any;
  projects: AdminPanelRelationPropertyModification<Project_Plain>;
  Slug: string;
}